+++ packages/thumbprint-react/components/ModalCurtain/test.tsx
@@ -43,13 +43,16 @@ describe('ModalCurtain', () => {
     });
 
     test('`children` receives `curtainOnClick` and `curtainClassName`', () => {
-        let obj;
+        let obj: {
+            curtainOnClick?: (event: React.MouseEvent<HTMLElement, MouseEvent>) => void;
+            curtainClassName?: string;
+        } = {};
 
         mount(
             <ModalCurtain onCloseClick={jest.fn}>
-                {args => {
+                {(args): JSX.Element => {
                     obj = args;
-                    return null;
+                    return <div />;
                 }}
             </ModalCurtain>,
         );
@@ -63,7 +66,7 @@ describe('ModalCurtain', () => {
 
         const wrapper = mount(
             <ModalCurtain onCloseClick={onClick}>
-                {({ curtainOnClick }) => (
+                {({ curtainOnClick }): JSX.Element => (
                     <button data-test-id="goose" onClick={curtainOnClick} type="button" />
                 )}
             </ModalCurtain>,
@@ -80,7 +83,9 @@ describe('ModalCurtain', () => {
         const wrapper = mount(
             <div>
                 <button type="button" />
-                <ModalCurtain onCloseClick={jest.fn}>{() => 'Goose'}</ModalCurtain>
+                <ModalCurtain onCloseClick={jest.fn}>
+                    {(): JSX.Element => <div>Goose</div>}
+                </ModalCurtain>
             </div>,
         );
 
@@ -94,7 +99,7 @@ describe('ModalCurtain', () => {
         const onCloseClick = jest.fn();
         const wrapper = mount(
             <ModalCurtain stage="entered" onCloseClick={onCloseClick}>
-                {() => (
+                {(): JSX.Element => (
                     <>
                         <button type="button" />
                         <input />
@@ -118,10 +123,38 @@ describe('ModalCurtain', () => {
 
         expect(onCloseClick).toHaveBeenCalledTimes(0);
 
-        document.dispatchEvent(new KeyboardEvent('keyup', { keyCode: ENTER_KEY }));
-        document.dispatchEvent(new KeyboardEvent('keyup', { keyCode: BACKSPACE_KEY }));
-        document.dispatchEvent(new KeyboardEvent('keyup', { keyCode: SPACE_KEY }));
-        document.dispatchEvent(new KeyboardEvent('keyup', { keyCode: LOWERCASE_A_KEY }));
+        // NOTE: it seems that Enzyme does not respond to the standard properties on
+        // `KeyboardEventInit`, either `code: 'Escape'` or `key: 'Escape'`. It only responds to
+        // the deprecated property `keyCode`, which is not in the type defintion. As such we
+        // have to cast the type here to prevent an error.
+        document.dispatchEvent(
+            new KeyboardEvent('keyup', {
+                key: 'Enter',
+                code: 'Enter',
+                keyCode: ENTER_KEY,
+            } as KeyboardEventInit),
+        );
+        document.dispatchEvent(
+            new KeyboardEvent('keyup', {
+                key: 'Backspace',
+                code: 'Backspace',
+                keyCode: BACKSPACE_KEY,
+            } as KeyboardEventInit),
+        );
+        document.dispatchEvent(
+            new KeyboardEvent('keyup', {
+                key: 'Space',
+                code: 'Space',
+                keyCode: SPACE_KEY,
+            } as KeyboardEventInit),
+        );
+        document.dispatchEvent(
+            new KeyboardEvent('keyup', {
+                key: 'a',
+                code: 'a',
+                keyCode: LOWERCASE_A_KEY,
+            } as KeyboardEventInit),
+        );
 
         expect(onCloseClick).toHaveBeenCalledTimes(0);
     });
@@ -133,7 +166,11 @@ describe('ModalCurtain', () => {
 
             expect(onCloseClick).toHaveBeenCalledTimes(0);
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(1);
@@ -147,7 +184,11 @@ describe('ModalCurtain', () => {
 
             expect(onCloseClick).toHaveBeenCalledTimes(0);
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(1);
@@ -177,7 +218,11 @@ describe('ModalCurtain', () => {
             const onCloseClick = jest.fn();
             mount(<ModalCurtain stage="exited" onCloseClick={onCloseClick} shouldCloseOnEscape />);
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(0);
@@ -193,7 +238,11 @@ describe('ModalCurtain', () => {
                 />,
             );
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(0);
@@ -209,7 +258,11 @@ describe('ModalCurtain', () => {
                 />,
             );
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(0);
@@ -221,7 +274,11 @@ describe('ModalCurtain', () => {
                 <ModalCurtain stage="entered" onCloseClick={onCloseClick} shouldCloseOnEscape />,
             );
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(1);
@@ -239,7 +296,11 @@ describe('ModalCurtain', () => {
                 <ModalCurtain stage="entered" onCloseClick={onCloseClick} shouldCloseOnEscape />,
             );
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(1);
@@ -257,7 +318,11 @@ describe('ModalCurtain', () => {
                 <ModalCurtain stage="entered" onCloseClick={onCloseClick} shouldCloseOnEscape />,
             );
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(1);
@@ -275,7 +340,11 @@ describe('ModalCurtain', () => {
                 <ModalCurtain stage="entered" onCloseClick={onCloseClick} shouldCloseOnEscape />,
             );
 
-            const event = new KeyboardEvent('keyup', { keyCode: ESC_KEY });
+            const event = new KeyboardEvent('keyup', {
+                key: 'Escape',
+                code: 'Escape',
+                keyCode: ESC_KEY,
+            } as KeyboardEventInit);
             document.dispatchEvent(event);
 
             expect(onCloseClick).toHaveBeenCalledTimes(1);