+++ packages/thumbprint-react/index.ts
@@ -28,7 +28,7 @@ export {
     ModalAnimatedWrapper,
 } from './components/Modal/index';
 export { default as ModalBase } from './components/ModalBase/index.jsx';
-export { default as ModalCurtain } from './components/ModalCurtain/index.jsx';
+export { default as ModalCurtain } from './components/ModalCurtain/index';
 export { default as ModalStandard } from './components/ModalStandard/index.jsx';
 export {
     default as Popover,