+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+-   [Patch] Rewrite `ModalCurtain` component in TypeScript.
+
 ## 12.3.0 - 2020-03-16
 
 ### Added